final: move files to correct folder and fix empty files
