--- weeks/week-02/solutions/1114405054/TEST_CASES.md
@@ -1,3 +1,0 @@
-# TEST_CASES.md
-- 已通過所有 9 個單元測試。
-- 包含正常情境、同分排序及空字串輸入測試。
