+++ weeks/week-02/solutions/1114405054/task2_student_ranking.py
@@ -1,0 +1,2 @@
+def get_ranking(students, k):
+    return sorted(students, key=lambda x: (-x[1], x[2], x[0]))[:k]
