+++ weeks/week-02/solutions/1114405054/tests/test_task1.py
@@ -1,0 +1,10 @@
+import unittest
+from task1_sequence_clean import clean_sequence
+class TestTask1(unittest.TestCase):
+    def test_basic(self):
+        res = clean_sequence("5 3 5 2")
+        self.assertEqual(res["dedupe"], [5, 3, 2])
+    def test_empty(self):
+        self.assertEqual(len(clean_sequence("")["dedupe"]), 0)
+    def test_evens(self):
+        self.assertEqual(clean_sequence("1 3 5")["evens"], [])
